Consolidate code, add all()
